+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 All Notable changes to `flipboxdigital/hubspot` will be documented in this file
 
+## Unreleased
+### Added
+- Contact batch
+- Contact List add/remove
+
+
 ## 1.0.0-beta - 2017-6-13
 ### Added
 - Initial release!